[RENAME] 기존 component 폴더 내 meal-plan 폴더 안 파일을 chip 폴더로 분리
